NEW FILES
Initial commit set up the project structure with HTML and CSS files and established the basic blog layout. Added blog posts for "Souls on the Banks of the Acheron," "Lucifer," and "The Voyage of Life," including detailed descriptions and author information. Added "Read more" links to Wikipedia for additional information on each artwork. Expanded categories and tags to cover new art styles and genres, including music and literature. Enhanced search functionality and added new sections to the sidebar, such as categories, tags, and events. Revised the "About" section to better describe the blog's focus on various art forms. Included a footer with copyright information for 2024. Made visual improvements to blog posts, navigation, and sidebar elements.